refactor: modularize monolithic EventManager IPC handlers into domain-specific classes
