+++ src/browser/main/ipc/handlers/PageContentHandler.ts
@@ -1,0 +1,40 @@
+import { ipcMain } from "electron";
+import { BaseHandler } from "./BaseHandler";
+
+export class PageContentHandler extends BaseHandler {
+  public register(): void {
+    // Get page content
+    ipcMain.handle("get-page-content", async () => {
+      if (this.mainWindow.activeTab) {
+        try {
+          return await this.mainWindow.activeTab.getTabHtml();
+        } catch (error) {
+          console.error("Error getting page content:", error);
+          return null;
+        }
+      }
+      return null;
+    });
+
+    // Get page text
+    ipcMain.handle("get-page-text", async () => {
+      if (this.mainWindow.activeTab) {
+        try {
+          return await this.mainWindow.activeTab.getTabText();
+        } catch (error) {
+          console.error("Error getting page text:", error);
+          return null;
+        }
+      }
+      return null;
+    });
+
+    // Get current URL
+    ipcMain.handle("get-current-url", () => {
+      if (this.mainWindow.activeTab) {
+        return this.mainWindow.activeTab.url;
+      }
+      return null;
+    });
+  }
+}
