+++ src/browser/main/ipc/handlers/SidebarHandler.ts
@@ -1,0 +1,30 @@
+import { ipcMain } from "electron";
+import { BaseHandler } from "./BaseHandler";
+
+export class SidebarHandler extends BaseHandler {
+  public register(): void {
+    // Toggle sidebar
+    ipcMain.handle("toggle-sidebar", () => {
+      this.mainWindow.sidebar.toggle();
+      this.mainWindow.updateAllBounds();
+      return true;
+    });
+
+    // Chat message
+    ipcMain.handle("sidebar-chat-message", async (_, request) => {
+      // The LLMClient handles getting the screenshot and context directly
+      await this.mainWindow.sidebar.client.sendChatMessage(request);
+    });
+
+    // Clear chat
+    ipcMain.handle("sidebar-clear-chat", () => {
+      this.mainWindow.sidebar.client.clearMessages();
+      return true;
+    });
+
+    // Get messages
+    ipcMain.handle("sidebar-get-messages", () => {
+      return this.mainWindow.sidebar.client.getMessages();
+    });
+  }
+}
