+++ src/browser/main/ipc/handlers/TabHandler.ts
@@ -1,0 +1,136 @@
+import { ipcMain } from "electron";
+import { BaseHandler } from "./BaseHandler";
+
+export class TabHandler extends BaseHandler {
+  public register(): void {
+    // Handle topbar dynamic height adjustments
+    ipcMain.handle("set-topbar-height", (_, height: number) => {
+      if (this.mainWindow.topBar) {
+        this.mainWindow.topBar.setHeight(height);
+      }
+    });
+
+    // Create new tab
+    ipcMain.handle("create-tab", (_, url?: string) => {
+      const newTab = this.mainWindow.createTab(url);
+      return { id: newTab.id, title: newTab.title, url: newTab.url };
+    });
+
+    // Close tab
+    ipcMain.handle("close-tab", (_, id: string) => {
+      this.mainWindow.closeTab(id);
+    });
+
+    // Switch tab
+    ipcMain.handle("switch-tab", (_, id: string) => {
+      this.mainWindow.switchActiveTab(id);
+    });
+
+    // Get tabs
+    ipcMain.handle("get-tabs", () => {
+      const activeTabId = this.mainWindow.activeTab?.id;
+      return this.mainWindow.allTabs.map((tab) => ({
+        id: tab.id,
+        isActive: activeTabId === tab.id,
+        title: tab.title,
+        url: tab.url,
+        canGoBack: tab.webContents.navigationHistory.canGoBack(),
+        canGoForward: tab.webContents.navigationHistory.canGoForward(),
+      }));
+    });
+
+    // Navigation (for compatibility with existing code)
+    ipcMain.handle("navigate-to", (_, url: string) => {
+      if (this.mainWindow.activeTab) {
+        void this.mainWindow.activeTab.loadURL(url);
+      }
+    });
+
+    ipcMain.handle("navigate-tab", async (_, tabId: string, url: string) => {
+      const tab = this.mainWindow.getTab(tabId);
+      if (tab) {
+        await tab.loadURL(url);
+        return true;
+      }
+      return false;
+    });
+
+    ipcMain.handle("go-back", () => {
+      if (this.mainWindow.activeTab) {
+        this.mainWindow.activeTab.goBack();
+      }
+    });
+
+    ipcMain.handle("go-forward", () => {
+      if (this.mainWindow.activeTab) {
+        this.mainWindow.activeTab.goForward();
+      }
+    });
+
+    ipcMain.handle("reload", () => {
+      if (this.mainWindow.activeTab) {
+        this.mainWindow.activeTab.reload();
+      }
+    });
+
+    // Tab-specific navigation handlers
+    ipcMain.handle("tab-go-back", (_, tabId: string) => {
+      const tab = this.mainWindow.getTab(tabId);
+      if (tab) {
+        tab.goBack();
+        return true;
+      }
+      return false;
+    });
+
+    ipcMain.handle("tab-go-forward", (_, tabId: string) => {
+      const tab = this.mainWindow.getTab(tabId);
+      if (tab) {
+        tab.goForward();
+        return true;
+      }
+      return false;
+    });
+
+    ipcMain.handle("tab-reload", (_, tabId: string) => {
+      const tab = this.mainWindow.getTab(tabId);
+      if (tab) {
+        tab.reload();
+        return true;
+      }
+      return false;
+    });
+
+    ipcMain.handle("tab-screenshot", async (_, tabId: string) => {
+      const tab = this.mainWindow.getTab(tabId);
+      if (tab) {
+        const image = await tab.screenshot();
+        return image.toDataURL();
+      }
+      return null;
+    });
+
+    ipcMain.handle("tab-run-js", async (_, tabId: string, code: string) => {
+      const tab = this.mainWindow.getTab(tabId);
+      if (tab) {
+        return tab.runJs(code);
+      }
+      return null;
+    });
+
+    // Tab info
+    ipcMain.handle("get-active-tab-info", () => {
+      const { activeTab } = this.mainWindow;
+      if (activeTab) {
+        return {
+          canGoBack: activeTab.webContents.navigationHistory.canGoBack(),
+          canGoForward: activeTab.webContents.navigationHistory.canGoForward(),
+          id: activeTab.id,
+          title: activeTab.title,
+          url: activeTab.url,
+        };
+      }
+      return null;
+    });
+  }
+}
